+++ cli/internal/api/client.go
@@ -25,3 +25,52 @@ func NewClient(token string) (*cloudquery_api.ClientWithResponses, error) {
 	}
 	return c, nil
 }
+
+func NewAnonymousClient() (*cloudquery_api.ClientWithResponses, error) {
+	c, err := cloudquery_api.NewClientWithResponses(env.GetEnvOrDefault(envAPIURL, defaultAPIURL))
+	if err != nil {
+		return nil, fmt.Errorf("failed to create api client: %w", err)
+	}
+	return c, nil
+}
+
+func ListAllPlugins(cl *cloudquery_api.ClientWithResponses) ([]cloudquery_api.ListPlugin, error) {
+	page := cloudquery_api.Page(1)
+	perPage := cloudquery_api.PerPage(100)
+	plugins := make([]cloudquery_api.ListPlugin, 0)
+	for {
+		resp, err := cl.ListPluginsWithResponse(context.Background(), &cloudquery_api.ListPluginsParams{
+			PerPage: &perPage,
+			Page:    &page,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("failed to list plugins: %w", err)
+		}
+		if resp.JSON200 == nil {
+			return nil, fmt.Errorf("failed to list plugins: %w", err)
+		}
+		if resp.StatusCode() != http.StatusOK || resp.JSON200 == nil {
+			return nil, fmt.Errorf("failed to list plugins: %s", resp.Status())
+		}
+		plugins = append(plugins, resp.JSON200.Items...)
+		if resp.JSON200.Metadata.LastPage == nil || *resp.JSON200.Metadata.LastPage <= int(page) {
+			break
+		}
+		page++
+	}
+	return plugins, nil
+}
+
+func GetPluginVersion(cl *cloudquery_api.ClientWithResponses, teamName string, kind cloudquery_api.PluginKind, pluginName, pluginVersion string) (*cloudquery_api.PluginVersionDetails, error) {
+	resp, err := cl.GetPluginVersionWithResponse(context.Background(), teamName, kind, pluginName, pluginVersion)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get plugin version %s/%s@%s: %w", teamName, pluginName, pluginVersion, err)
+	}
+	if resp.JSON200 == nil {
+		return nil, fmt.Errorf("failed to get plugin version %s/%s@%s: %w", teamName, pluginName, pluginVersion, err)
+	}
+	if resp.StatusCode() != http.StatusOK {
+		return nil, fmt.Errorf("failed to get plugin version %s/%s@%s: %s", teamName, pluginName, pluginVersion, resp.Status())
+	}
+	return resp.JSON200, nil
+}