+++ plugins/destination/mysql/cloud-config-ui/e2e-tests/validate-config.spec.ts
@@ -1,9 +1,8 @@
 import test, { expect } from '@playwright/test';
-import fs from 'node:fs';
-import YAML from 'yaml';
 
 test('Submit the form', async ({ page }) => {
   await page.goto('/');
+
   await page.getByLabel('Host').click();
   await page.getByLabel('Host').fill('localhost');
   await page.getByLabel('Port').click();
@@ -31,68 +30,18 @@ test('Submit the form', async ({ page }) => {
   await page.getByLabel('Write Timeout').click();
   await page.getByLabel('Write Timeout').fill('8');
 
-  await page.getByLabel('Migrate mode *').click();
+  await page.getByLabel('Migrate mode').click();
   await page.getByRole('option', { name: 'forced' }).click();
-  await page.getByLabel('Write mode *').click();
+  await page.getByLabel('Write mode').click();
   await page.getByRole('option', { name: 'overwrite', exact: true }).click();
 
   await page.getByRole('button', { name: 'Advanced Sync Options' }).click();
+  await page.getByLabel('Batch size', { exact: true }).click();
+  await page.getByLabel('Batch size', { exact: true }).fill('12');
+  await page.getByLabel('Batch size (bytes)').click();
+  await page.getByLabel('Batch size (bytes)').fill('2500');
 
-  await page.getByLabel('Batch size *', { exact: true }).click();
-  await page.getByLabel('Batch size *', { exact: true }).fill('12');
-  await page.getByLabel('Batch size (bytes) *').click();
-  await page.getByLabel('Batch size (bytes) *').fill('2500');
+  await page.getByRole('button', { name: 'Test connection' }).click();
 
-  await page.getByRole('button', { name: 'Submit' }).click();
-
-  const valuesText = await page
-    .locator('text=Values:')
-    .locator('xpath=following-sibling::*[1]')
-    .textContent();
-  expect(valuesText).toBeTruthy();
-
-  const spec = JSON.parse(valuesText as string);
-  expect(spec.spec.connection_string).toBe(
-    'john_doe:${password}@tcp(localhost:3306)/sample_db?tlsMode=preferred&parseTime=True&charset=utf8mb4&loc=Local&timeout=6s&readTimeout=7s&writeTimeout=8s',
-  );
-
-  if (process.env.E2E_TESTS_GENERATE_CONFIG === 'true') {
-    const destinationConfig = YAML.stringify({
-      kind: 'destination',
-      spec: {
-        name: 'mysql',
-        registry: 'local',
-        path: '../mysql',
-        spec: spec.spec,
-        write_mode: spec.writeMode,
-        migrate_mode: spec.migrateMode,
-      },
-    });
-
-    const sourceConfig = YAML.stringify({
-      kind: 'source',
-      spec: {
-        name: 'postgresql',
-        path: 'cloudquery/postgresql',
-        registry: 'cloudquery',
-        version: 'v6.2.5',
-        destinations: ['mysql'],
-        spec: {
-          connection_string: 'test',
-        },
-        tables: ['*'],
-      },
-    });
-
-    if (!fs.existsSync('temp')) {
-      fs.mkdirSync('temp');
-    }
-
-    fs.writeFileSync('./temp/config.yml', `${sourceConfig}---\n${destinationConfig}`);
-
-    fs.writeFileSync(
-      './temp/.env',
-      `${spec.envs.map((env: { name: string; value: string }) => `${env.name}=${env.value}`).join('\n')}`,
-    );
-  }
+  await expect(page.getByText('Testing the destination connection')).toBeVisible();
 });