+++ plugins/source/hackernews/cloud-config-ui/src/utils/prepareSubmitValues.ts
@@ -1,23 +1,19 @@
 import { PluginUiMessagePayload } from '@cloudquery/plugin-config-ui-connector';
-
-import { FormValues } from './formSchema';
+import { corePrepareSubmitValues, PluginTable } from '@cloudquery/plugin-config-ui-lib';
 
 export function prepareSubmitValues(
-  values: FormValues,
+  values: Record<string, any>,
+  tablesList?: PluginTable[],
 ): PluginUiMessagePayload['validation_passed']['values'] {
-  const envs = [] as Array<{ name: string; value: string }>;
+  const payload = corePrepareSubmitValues(values, tablesList);
 
-  const spec = {
-    start_time: values.spec.startTimeEnabled ? values.spec.startTime.toISOString() : undefined,
-    item_concurrency: values.spec.itemConcurrency,
-  };
+  if (values.item_concurrency) {
+    payload.spec.item_concurrency = Number(values.item_concurrency);
+  }
 
-  return {
-    name: values.name,
-    envs,
-    tables: Object.keys(values.tables).filter(
-      (key) => values.tables[key as keyof typeof values.tables],
-    ),
-    spec,
-  };
+  if (values._startTimeEnabled) {
+    payload.spec.start_time = values.start_time.toISOString();
+  }
+
+  return payload;
 }