+++ plugins/destination/gremlin/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [2.5.13](https://github.com/cloudquery/cloudquery/compare/plugins-destination-gremlin-v2.5.12...plugins-destination-gremlin-v2.5.13) (2024-09-05)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/config to v1.27.31 ([#19043](https://github.com/cloudquery/cloudquery/issues/19043)) ([3854601](https://github.com/cloudquery/cloudquery/commit/3854601e7decd3f3ab55f384f926f24717cf163e))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.61.0 ([#19069](https://github.com/cloudquery/cloudquery/issues/19069)) ([bf8ab33](https://github.com/cloudquery/cloudquery/commit/bf8ab3356a44e3fee8e03f68fbc3994471cdb6fa))
+
+## [2.5.12](https://github.com/cloudquery/cloudquery/compare/plugins-destination-gremlin-v2.5.11...plugins-destination-gremlin-v2.5.12) (2024-08-27)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/aws/aws-sdk-go-v2/config to v1.27.30 ([#19004](https://github.com/cloudquery/cloudquery/issues/19004)) ([92c1b1c](https://github.com/cloudquery/cloudquery/commit/92c1b1ce773ba5d1b00e45b33f71c27f63bde81f))
+
 ## [2.5.11](https://github.com/cloudquery/cloudquery/compare/plugins-destination-gremlin-v2.5.10...plugins-destination-gremlin-v2.5.11) (2024-08-20)